+++ Sources/Sentry/SentrySDK.m
@@ -42,6 +42,7 @@
 
 #if SENTRY_TARGET_PROFILING_SUPPORTED
 #    import "SentryContinuousProfiler.h"
+#    import "SentryProfileConfiguration.h"
 #    import "SentryProfiler+Private.h"
 #endif // SENTRY_TARGET_PROFILING_SUPPORTED
 
@@ -669,7 +670,8 @@ static NSDate *_Nullable startTimestamp = nil;
             return;
         }
 
-        if (sentry_profilerSessionSampleDecision.decision != kSentrySampleDecisionYes) {
+        if (sentry_profileConfiguration.profilerSessionSampleDecision.decision
+            != kSentrySampleDecisionYes) {
             SENTRY_LOG_DEBUG(
                 @"The profiling session has been sampled out, no profiling will take place.");
             return;