+++ Sources/Sentry/SentryTimeToDisplayTracker.m
@@ -21,6 +21,7 @@
 
 #    if SENTRY_TARGET_PROFILING_SUPPORTED
 #        import "SentryLaunchProfiling.h"
+#        import "SentryProfiler+Private.h"
 #    endif // SENTRY_TARGET_PROFILING_SUPPORTED
 
 @interface SentryTimeToDisplayTracker () <SentryFramesTrackerListener>
@@ -175,8 +176,8 @@
         if (!_waitForFullDisplay) {
             [SentryDependencyContainer.sharedInstance.framesTracker removeListener:self];
 #    if SENTRY_TARGET_PROFILING_SUPPORTED
-            if ([SentrySDK.options isProfilingCorrelatedToTraces]) {
-                sentry_stopAndDiscardLaunchProfileTracer();
+            if (sentry_isLaunchProfileCorrelatedToTraces()) {
+                sentry_stopAndDiscardLaunchProfileTracer(SentrySDK.currentHub);
             }
 #    endif // SENTRY_TARGET_PROFILING_SUPPORTED
         }
@@ -187,8 +188,8 @@
         self.fullDisplaySpan.timestamp = newFrameDate;
         [self.fullDisplaySpan finish];
 #    if SENTRY_TARGET_PROFILING_SUPPORTED
-        if ([SentrySDK.options isProfilingCorrelatedToTraces]) {
-            sentry_stopAndDiscardLaunchProfileTracer();
+        if (sentry_isLaunchProfileCorrelatedToTraces()) {
+            sentry_stopAndDiscardLaunchProfileTracer(SentrySDK.currentHub);
         }
 #    endif // SENTRY_TARGET_PROFILING_SUPPORTED
     }