+++ Tests/SentryTests/SentryContinuousProfiler+Test.h
@@ -10,6 +10,7 @@
 
 + (void)stopTimerAndCleanup;
 + (nullable SentryProfiler *)profiler;
++ (void)resetForTesting;
 
 @end